backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
psycopg[binary]>=3.2,<4.0
